eps using long

+++ eps2.cpp
@@ -3,8 +3,9 @@ using namespace std;
 int main()
 
 //this code will find the epsilon of machine
+//machine eps using long 
 {
-     double x = 2.0;
+    long double x = 2.0;
     int f;
     for (int  i = 1;   1+x>1; i ++)
     {
@@ -18,3 +19,4 @@ int main()
 
     return 0;
 }
+
